Move SPI code out of board_enable.c where it started its life
The SPI chip finding and SPI chip accessor code is moved as well. This can
be split later if we feel like it. The non-use of svn cp is intentional
because the only history we'd have to preserve are a few commits which were
early prototypes of chip identification code. For those who intend to look
at that history, they can look at board_enable.c revision 2853.

Corresponding to flashrom svn r145 and coreboot v2 svn r2857.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Stefan Reinauer <stepan@coresystems.de>