Add generic SPI flash erase and write support to flashrom
The first chip the code was tested and verified with is the Macronix MX25L4005,
but other chips should work as well. Timeouts are still hardcoded to data
sheet maxima, but the status register checking code is already there.
Thanks to Harald Gutmann for the initial code on which this is loosely based.

Corresponding to flashrom svn r152 and coreboot v2 svn r2874.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Stefan Reinauer <stepan@coresystems.de>