Fix ATMEL 29C020 detection with flashrom
The JEDEC probe routine had a delay of 10 us after entering ID mode and this
was insufficient for the 29C020. The data sheet claims we have to wait 10 ms,
but tests have shown that 20 us suffice. Allow for variations in chip delays
with a factor of 2 safety margin.

Corresponding to flashrom svn r159 and coreboot v2 svn r2962.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Peter Stuge <peter@stuge.se>