The generic jedec.c does not work for the ST M50FLW flas devices, because they need an unlock command first
For this reason, ST M50FLW support is moved to a new HW support module,
because any change in jedec.c would bear the risk to cause problems with the
already supported devices. It's already tested with ST M50FLW080A; the other
chips of this family i dont have available, so i couldnt test it.

Corresponding to flashrom svn r219 and coreboot v2 svn r3274.

Signed-off-by: Claus Gindhart <claus.gindhart@kontron.com>
Acked-by: Stefan Reinauer <stepan@coresystems.de>