Add support for the JEDEC RES (Read Electronic Signature and Resume fro Powerdown) SPI command to flashrom to identify older SPI chips which
Can't handle JEDEC RDID. Since RES gives a one-byte identifier which is shared
among many different vendors and even different sizes, we want to match RES
as a last resort if RDID returns 0xff 0xff 0xff. This is a heavily reworked
version of a patch by Fredrik Tolf, which was

Corresponding to flashrom svn r235 and coreboot v2 svn r3320.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Signed-off-by: Fredrik Tolf <fredrik@dolda2000.com>
Acked-by: Peter Stuge <peter@stuge.se>