The existing check in probe_spi_res() was right for SPI controller which support all commands, but may not exist
For controllers which support only a subset of commands, it will fail in
unexpected ways. Even if a command is supported by the controller, it may
be unavailable if the controller is locked down. The new logic checks if
RDID could be issued and its return values made sense (not 0xff 0xff 0xff).
In that case, RES probing is not performed. Otherwise, we try RES. There is
one drawback: If RDID returned unexpected values, we don't issue a RES probe.
However, in that case we should try to match RDID anyway.

Corresponding to flashrom svn r348 and coreboot v2 svn r3774.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: FENG yu ning <fengyuning1984@gmail.com>