All "unknown xy SPI chip" entries claim to have status UNTESTED fo probe/read/erase/write
That is incorrect. A bit of confusion comes from how the #defines are named.
We call them TEST_BAD_*, but the message printed by flashrom says: "This flash
part has status NOT WORKING for operations:" Something that is unimplemented
is definitely not working. Neither of the chip entries mentioned above has
erase or write functions implemented, so erase and write are not working.
Since their size is unknown, we can't read them in. That means read is not
working as well. Probing is a different matter. If a chip-specific probe
function had matched, we wouldn't have to handle the chip with the "unknown xy
SPI chip" fallback. I'm tempted to call that "not working" as well, but I'm
open to discussion on this point.

Corresponding to flashrom svn r439 and coreboot v2 svn r4177.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Uwe Hermann <uwe@hermann-uwe.de>