The flashrom makefile wants to redirect both stdout and stderr t /dev/null for one compile test
The old variant of using &>/dev/null works on bash and zsh, but not on dash
and tcsh. dash and tcsh interpret it as "background command and truncate
/dev/null" which is not what we want. >& works on tcsh and bash, but it is not
POSIX compliant. Since make uses /bin/sh and /bin/sh has to be POSIX compliant,
we can use the POSIX variant of stderr and stdout redirection. >/dev/null 2>&1
is POSIX compliant. This is specified in SuSv3, Shell Command Language,
sections 2.7.2 and 2.7.6.
Signed-off-by: Stephan Guilloux <stephan.guilloux@free.fr>

Corresponding to flashrom svn r443 and coreboot v2 svn r4211.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Ward Vandewege <ward@gnu.org>