Flashrom assumes that the flash chip contents are available via mmap i no read function is defined
This special case is handled in lots of places all over the code. Remove the
special case and use the read_memmapped function. Not only does this allow us
to fix a read bug in flashrom I recently uncovered on ICH SPI, it also allows
us to add support for Paraflasher to flashrom.

Corresponding to flashrom svn r473.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Uwe Hermann <uwe@hermann-uwe.de>