Until the ICH SPI driver can handle preopcodes as standalone opcodes, w should handle such special opcode failure gracefully on ICH and
Compatible chipsets. This fixes status register writes on almost all
ICH+VIA SPI masters. The fix is almost identical to r484, but this time
it affects the EWSR (Enable Write Status Register) opcode instead of the
WREN (Write Enable) opcode. With the differentiated return codes introduced
in r500, the workaround is more precise this time. The old WREN workaround
was updated as well.

Corresponding to flashrom svn r514.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: FENG Yu Ning <fengyuning1984@gmail.com>
Acked-by: Cristi Magherusan <cristi.magherusan@net.utcluj.ro>