This patch introduces two new targets which are designed to make th life of packagers easier
In particular, it should no longer be necessary to patch the makefile for
hassle-free compilation. The targets are: make export make tarball Both
preserve svn revisions and the exported tree does not depend on subversion
in any way or shape. Documentation for this feature has been added to README.
We need this for five reasons: 1. Packagers currently have to patch flashrom
source to compile it on systems without subversion. We should make it easier
for them. 2. Snapshot tarballs currently have a .svn 1.5 directory included
but this will cause errors for users with older svn 1.4. Not requiring
subversion for snapshot compilation is best. 3. Since packagers seldom the
svn revision in their fixup patches, some packages out there have incorrect
or no revision, only major version numbers. 4. Releasing a new version of
flashrom needs too many changes to the makefile which have to be reverted
instantly after the release. That is unnecessary churn. 5. Making a release
is easy with the change. Update the major version, then run "make tarball".

Corresponding to flashrom svn r587.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>  
Acked-by: Uwe Hermann <uwe@hermann-uwe.de>