Split hardware access, OS abstraction and chip drivers out of flash.h t get a better overview of what belongs where
This patch is only the first step, but it hopefully will make working with the
code and especially porting to new platforms easier. Subsequent patches should
move #includes for the newly created files hwaccess.h and chipdrivers.h from
flash.h to the files which need them. Programmers should live in a separate
header file as well.

Corresponding to flashrom svn r799.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Tested-by: Idwer Vollering <vidwer@gmail.com>
Acked-by: Sean Nelson <audiohacked@gmail.com>