SuperI/O detection now happens unconditionally and before the chipse enable
We could run it after chipset enable, but it definitely has to happen before
board enable because the board enable usually accesses the SuperI/O. With this
patch, it is possible to add a struct superio to the board enable table for
more accurate matching in case subsystem IDs are ambiguous. This patch focuses
on the generic infrastructure aspect and on support for IT8712F/IT8716F.
Thanks go to Adrian Glaubitz and Ward Vandewege for testing.

Corresponding to flashrom svn r813.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Luc Verhaegen <libv@skynet.be>
Acked-by: Adrian Glaubitz <glaubitz@physik.fu-berlin.de>