Create a physical memory mapping function which requests cached readonl memory
This should take care of picky Linux kernels which do not allow uncached
mappings to cached areas. Handle mapping failure gracefully (no forced exit())
if the caller specifies it. Such cached areas which can handle mapping failure
are DMI tables and coreboot tables. On failure we just ignore those tables.
That is not perfect, but a lot better than aborting flashrom due to an error
in nonessential functionality. This should fix flashrom on a sizable number
of machines where it currently aborts early. Yes, I could have exploited
a Linux kernel bug to "solve" this, but relying on such bugs is not exactly
the best idea.

Corresponding to flashrom svn r889.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Vincent Pelletier <plr.vincent@gmail.com>