SPI RES is the most unreliable way to identify chips because it onl returns a 1-byte ID for most chips
For every given ID out there, probably a dozen incompatible flash chips
match it. We already refuse to identify a chip with RES if that chip responds
to RDID (3 bytes, good match), and with this patch we additionally refuse RES
if the chip responds to REMS (2 bytes, still a good match). This increases
matching accuracy a lot. Besides that, the RDID/REMS response checking has
been cleaned up for better readability.

Corresponding to flashrom svn r899.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Sean Nelson <audiohacked@gmail.com>