Idwer Vollering reported problems with the current libpci check o FreeBSD 8.0-RELEASE i386
This is caused by a strict linker. Parsing linker error messages is an exercise
in futility, and library detection with $CC--print-file-name is totally
useless for libraries outside the standard hardcoded builtin gcc search path
(probably the same for other compilers as well). Look for libpciutils instead
of libpci on NetBSD during the check.

Corresponding to flashrom svn r968.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Idwer Vollering <vidwer@gmail.com>