Clean up manufacturer mainboard links
The Asus A8NE-FM/S does exist, you find original ASUS pdf manuals in the
internet, but seems to be an OEM board that is not documented on the
Asus page.

A lot of MSI boards are OEM boards that have no page at the MSI site...

This patch also adds links in the Wiki for boards from the board enable
table.

Corresponding to flashrom svn r976.

Signed-off-by: Michael Karcher <flashrom@mkarcher.dialup.fu-berlin.de>
Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>