Fix Tyan S2915 OEM board by commenting out MCP55 LPC bridge PCI I 10de:0361 which is the secondary LPC bridge
The same effect could be achieved by refusing to run enable_flash_mcp55 if
the device class is not ISA bridge [0601]. Thanks to Alessandro Polverini,
Joel Robertson, Nicolas Aveline, Phil LoCascio and Nils-Helge Garli Hegvik
for testing flashrom on hardware and Michael Karcher for analyzing the
factory BIOS for clues. In the end, no board enable was needed and it was a
pure chipset issue.

Corresponding to flashrom svn r1007.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Michael Karcher <flashrom@mkarcher.dialup.fu-berlin.de>