Add support for the Amontec JTAGkey2, se http://www.amontec.com/jtagkey2.shtml
http://www.amontec.com/jtagkey.shtml This FTDI 2232H variant has an additional
output enable, which will be set to its "on" (L) when CS is pulled low.
But it lacks a power supply and you need an external 3.3V source. The attached
patch adds "jtagkey" as "type" parameter for ft2232_spi. It should work with
all JTAGkeys (JTAGkey, JTAGkey-tiny and JTAGkey2) but I only have a JTAGkey2
here for testing. Add all FT2232H/FT4232H based programmers to the list
printed with flashrom -L

Corresponding to flashrom svn r1119.

Signed-off-by: Jörg Fischer <turboj@gmx.de>
Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>