Speed up Bus Pirate by exploiting the FTDI FIFO
If we violate the raw SPI communication protocol requirements of the
Bus Pirate (namely, waiting for the completion of one command before sending
the next one), we can reduce the number of round trips by a factor of 3.
The FT2232 chip present in the Bus Pirate has a big enough buffer (at least
128 bytes IIRC) to avoid overflows in the tiny buffer of the Bus Pirate PIC.
Thanks to Daniel Flinkmann for sponsoring development of this patch.
Tested-by: Daniel Flinkmann <DFlinkmann@gmx.de>

Corresponding to flashrom svn r1120.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Daniel Flinkmann <dflinkmann@gmx.de>