Update the ITE IT8500 EC support to match the current state of th flashrom-chromium tree
This code has been deployed and tested to work on the Cr-48. There are a
few caveats, though: - The boot BIOS straps register must be modified to
select LPC. This can be done with the "select_bbs.sh" script (Install iotools
at http://code.google.com/p/iotools/ before using select_bbs). - It is very
important to disable power management daemons before running flashrom on
this EC. I commented out the brute force method we use in the Chromium OS
branch that disables powerd, since IIRC Carl-Daniel has a better approach in
the works. - Due to dependencies which may be introduced by the OEM/ODM EC
firmware, the code is not guaranteed to work for anything other than the
Cr-48. Carl-Daniel comments: Code is not hooked up yet because probing needs
to be sorted out.

Corresponding to flashrom svn r1263.

Signed-off-by: David Hendricks <dhendrix@google.com>
Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>