Add board enable for ASUS P5LD2-VM DH
Tested on real hardware by TeslaBIOS.
Besides the usual board_enable stuff the GPIO definitions for the ICH7 DH
were also missing. Apparently Intel forgot to add the PCI IDs for the desktop
version in the spec update, but the normal datasheet mentions the DH
desktop version so this should be fine...

Corresponding to flashrom svn r1764.

Signed-off-by: Stefan Tauner <stefan.tauner@alumni.tuwien.ac.at>
Acked-by: Stefan Tauner <stefan.tauner@alumni.tuwien.ac.at>